refactor LazyVideo to not have base class
